requirements.txt
numpy==1.17.2              
opencv-python==4.1.0.25            
plyfile==0.7.1               
progressbar==2.5                 
pyvista==0.23.0              
scipy==1.3.1               
tensorflow-gpu==2.0.0